+++ clients/admin-ui/src/features/plus/plus.slice.ts
@@ -26,6 +26,7 @@ import {
   ClassifySystem,
   CloudConfig,
   ConnectionConfigurationResponse,
+  ConnectionDatahubSyncResponse,
   ConsentableItem,
   CustomAssetType,
   CustomFieldDefinition,
@@ -474,6 +475,16 @@ const plusApi = baseApi.injectEndpoints({
       }),
       invalidatesTags: ["Consentable Items"],
     }),
+    syncDatahubConnection: build.mutation<
+      ConnectionDatahubSyncResponse,
+      { connectionKey: string; datasetIds: string[] }
+    >({
+      query: ({ connectionKey, datasetIds }) => ({
+        url: `plus/connection/datahub/${connectionKey}/sync`,
+        method: "POST",
+        body: { dataset_ids: datasetIds },
+      }),
+    }),
   }),
 });
 
@@ -515,6 +526,7 @@ export const {
   usePatchTcfPurposeOverridesMutation,
   useGetConsentableItemsQuery,
   useUpdateConsentableItemsMutation,
+  useSyncDatahubConnectionMutation,
 } = plusApi;
 
 export const selectHealth: (state: RootState) => HealthCheck | undefined =